Create loopings.sol
Let's put our new knowledge into action!.

We'll do a little lab where you will do a stress test a for loop. This is not graded but rather a way for you to practice ahead of our final challenge. Use this to try loops out and share what you learn. To get an easier result, you may want to reduce your gas limit from 3 million down to something like 300k in the deploy section of remix.

For this challenge, do a simple operation, and run it in 1000 loops.

Alternatively, try to do a complex operation and run it in 100 loops.

Does the transaction throw an error? Why?